Refactoring server to make twitch/youtube notifications modular
